requirements.txt
Pillow==9.5.0
Pillow==7.0.0
Pillow==10.0.0
rembg==2.0.49
streamlit==1.23.1
